backend/requirements.txt
asgiref==3.3.1
Django==3.1.13
djangorestframework==3.12.4
django-crispy-forms==1.11.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.1
django-multiselectfield==0.1.12